Create HW1stMarch
program:1 
The program is about if and else  by using try and except and we are giving the input as month number is given by user and the result will be the month name 
program :2 
The program is  to determine bill amount using match and cases.  we are giving input as units and the resultant will be the cost of that particular units
program: 3
The program is to print each element of list using for loop
program: 4
The program is to print each character of string 'Hyd' using for loop
program: 5
The program is to print each element of range using for loop
program: 6
The program is to find the output using for loop with dictionary using .keys we are getting the result as key values of dictionary
program: 7
The program is to  find the output using for loop with dictionary using .values we are getting the result as values 
program: 8
The program is to  find the output using for loop with dictionary using .items we are getting the result as ( key, values)
program:9
 The program is to  find the output using for loop with dictionary we are getting the result as ( key, values)
program : 10
The program is to find output using for loop with dictionary using .items we are getting the result as ( key, values)and the next two for loops cannot be unpack as they are int object.
program : 11
The program is to find output using for loop with integers where it gives error as for loop always allows and only for sequence objects only so int is non-sequence and it cant be iterable
program : 12
The program is to find output using for loop with empty sequences( tuple, list, set, dictionary and string) the result will be nothing and empty range gives error range(10,10) will return nothing. for non-sequence it throws the error.
program: 13
The program is about Nested  Loop of for loop the result will be  iteration of for loop and at the every end of the for loop will get a text as hello and after the end of the program will get bye.